move code for docker image building
